+++ core/src/main/java/net/cofcool/chaos/server/core/support/GlobalHandlerExceptionResolver.java
@@ -16,11 +16,7 @@
 
 package net.cofcool.chaos.server.core.support;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.OutputStream;
-import java.io.PrintStream;
 import java.lang.reflect.UndeclaredThrowableException;
 import java.util.NoSuchElementException;
 import javax.servlet.http.HttpServletRequest;
@@ -33,8 +29,11 @@ import net.cofcool.chaos.server.common.core.ServiceException;
 import net.cofcool.chaos.server.common.util.WebUtils;
 import net.cofcool.chaos.server.core.aop.ValidateInterceptor;
 import org.springframework.beans.factory.InitializingBean;
+import org.springframework.boot.autoconfigure.http.HttpMessageConverters;
 import org.springframework.core.Ordered;
+import org.springframework.http.MediaType;
 import org.springframework.http.converter.HttpMessageNotReadableException;
+import org.springframework.http.converter.HttpMessageNotWritableException;
 import org.springframework.util.Assert;
 import org.springframework.util.ClassUtils;
 import org.springframework.web.bind.MethodArgumentNotValidException;
@@ -44,7 +43,7 @@ import org.springframework.web.servlet.handler.AbstractHandlerExceptionResolver;
 import org.springframework.web.servlet.mvc.support.DefaultHandlerExceptionResolver;
 
 /**
- * 异常处理器, 应用处于 {@link ConfigurationSupport#isDebug()} 时优先级低于Spring默认异常解析器的, 其它情况优先级最高,
+ * 异常处理器, 把异常信息转换为 {@link Message}, "Content-Type" 为 "JSON", 应用处于 {@link ConfigurationSupport#isDebug()} 时优先级低于Spring默认异常解析器的, 其它情况优先级最高,
  * 部分 {@linkplain Message 描述信息} 通过 {@link ConfigurationSupport#getMessageWithKey(String, String, Object)} 创建,
  * 如 {@link ServiceException} 等
  *
@@ -56,6 +55,8 @@ import org.springframework.web.servlet.mvc.support.DefaultHandlerExceptionResolv
  * @author CofCool
  */
 public class GlobalHandlerExceptionResolver extends AbstractHandlerExceptionResolver implements InitializingBean {
+
+    protected final MediaType jsonType = MediaType.APPLICATION_JSON;
 
     private final HandlerExceptionResolver defaultExceptionResolver;
 
@@ -72,7 +73,7 @@ public class GlobalHandlerExceptionResolver extends AbstractHandlerExceptionReso
         }
     }
 
-    private ObjectMapper jacksonObjectMapper;
+    private HttpMessageConverters httpMessageConverters;
 
     private ConfigurationSupport configuration;
 
@@ -88,12 +89,12 @@ public class GlobalHandlerExceptionResolver extends AbstractHandlerExceptionReso
         defaultExceptionResolver = new DefaultHandlerExceptionResolver();
     }
 
-    protected ObjectMapper getJacksonObjectMapper() {
-        return jacksonObjectMapper;
+    protected HttpMessageConverters getHttpMessageConverters() {
+        return httpMessageConverters;
     }
 
-    public void setJacksonObjectMapper(ObjectMapper jacksonObjectMapper) {
-        this.jacksonObjectMapper = jacksonObjectMapper;
+    public void setHttpMessageConverters(HttpMessageConverters httpMessageConverters) {
+        this.httpMessageConverters = httpMessageConverters;
     }
 
     protected HandlerExceptionResolver getDefaultExceptionResolver() {
@@ -103,12 +104,12 @@ public class GlobalHandlerExceptionResolver extends AbstractHandlerExceptionReso
     @Override
     protected ModelAndView doResolveException(HttpServletRequest request,
                                               HttpServletResponse response, Object handler, Exception ex) {
-        printExceptionLog(request, handler, ex);
-
         Exception throwable = ex;
         if (ex instanceof UndeclaredThrowableException) {
             throwable = (Exception) ((UndeclaredThrowableException) ex).getUndeclaredThrowable();
         }
+
+        printExceptionLog(request, handler, throwable);
 
         if (throwable instanceof ServiceException) {
             return handleServiceException(response, (ServiceException) throwable);
@@ -123,7 +124,7 @@ public class GlobalHandlerExceptionResolver extends AbstractHandlerExceptionReso
         } else if (springDataAccessException != null && springDataAccessException.isAssignableFrom(throwable.getClass())) {
             return handleSqlException(response, throwable);
         } else {
-            return resolveOthersException(request, response, handler, ex);
+            return resolveOthersException(request, response, handler, throwable);
         }
     }
 
@@ -142,10 +143,7 @@ public class GlobalHandlerExceptionResolver extends AbstractHandlerExceptionReso
 
     protected void printExceptionLog(HttpServletRequest request, Object handler, Exception ex) {
         if (logger.isErrorEnabled() && ((!(ex instanceof ExceptionLevel)) || ((ExceptionLevel) ex).showable())) {
-            OutputStream outputStream = new ByteArrayOutputStream();
-            ex.printStackTrace(new PrintStream(outputStream));
-
-            logger.error("exception log: path=[" + WebUtils.getRealRequestPath(request) + "];exception=[" + outputStream.toString() + "]");
+            logger.error("exception log: path=[" + WebUtils.getRealRequestPath(request) + "]", ex);
         }
     }
 
@@ -235,15 +233,14 @@ public class GlobalHandlerExceptionResolver extends AbstractHandlerExceptionReso
         return EMPTY_MODEL_AND_VIEW;
     }
 
+    @SuppressWarnings("rawtypes")
     protected void writeMessage(HttpServletResponse response, Message message) {
         try {
             response.setCharacterEncoding("utf-8");
-            response.setHeader("Content-Type", "application/json;charset=UTF-8");
-            response.getWriter().append(jacksonObjectMapper.writeValueAsString(message));
-            response.flushBuffer();
-        } catch (IOException e) {
+            WebUtils.writeObjToResponse(getHttpMessageConverters(), response, message, MediaType.APPLICATION_JSON);
+        } catch (IOException | HttpMessageNotWritableException e) {
             if (logger.isErrorEnabled()) {
-                logger.error("parse exception error: ",e);
+                logger.error("parse exception error: ", e);
             }
         }
     }
@@ -251,6 +248,7 @@ public class GlobalHandlerExceptionResolver extends AbstractHandlerExceptionReso
     @Override
     public void afterPropertiesSet() throws Exception {
         Assert.notNull(configuration, "configuration must be specified");
+        Assert.notNull(httpMessageConverters, "httpMessageConverters must be specified");
 
         if (!configuration.isDebug()) {
             // 允许自定义更高优先级的异常处理器