+++ security-spring/src/main/java/net/cofcool/chaos/server/security/spring/authorization/JsonAuthenticationSuccessHandler.java
@@ -21,7 +21,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import net.cofcool.chaos.server.common.core.ConfigurationSupport;
-import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
+import org.springframework.boot.autoconfigure.http.HttpMessageConverters;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 
@@ -36,7 +36,7 @@ public class JsonAuthenticationSuccessHandler extends AbstractAuthenticationConf
 
     public JsonAuthenticationSuccessHandler(
         ConfigurationSupport configuration,
-        MappingJackson2HttpMessageConverter messageConverter) {
+        HttpMessageConverters messageConverter) {
         super(configuration, messageConverter);
     }